Fixed another bug in writeValue: valueOffset must be int64_t rather than just int. Two more programs pass.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@6258 1c3fd406-9068-4342-8353-511fcdf43622

+++ src/IValue.cpp
@@ -288,7 +288,7 @@ int IValue::setValue(int offset, int byte, uint8_t* content) {
 } 
 
 void IValue::writeValue(int offset, int byte, IValue* src) {
-  int valueOffset, length, bitOffset;
+  int64_t valueOffset, length, bitOffset;
   IValue* valueArray = static_cast<IValue*>(getIPtrValue());
 
   cout << "\t writing " << byte << " bytes\n" << endl;
